+++ elsa/cc_type.cc
@@ -1640,7 +1640,7 @@ inline unsigned cvHash(CVFlags cv)
 unsigned CVAtomicType::innerHashValue() const
 {
   // underlying atomic is pointer-based equality
-  return (unsigned)atomic + 
+  return (unsigned)(uintptr_t)atomic + 
          cvHash(cv);
          // T_ATOMIC is zero anyway
 }
@@ -2335,7 +2335,7 @@ unsigned PointerToMemberType::innerHashValue() const
   return atType->innerHashValue() * HASH_KICK +
          cvHash(cv) +
          T_POINTERTOMEMBER * TAG_KICK +
-         (unsigned)inClass();   // we'll see...
+         (unsigned)(uintptr_t)inClass();   // we'll see...
 }
